fix warning C4244
conversion from 'double' to 'int', possible loss of data

+++ systeminfo.cpp
@@ -5,7 +5,7 @@
 #include <string>
 #include <ctime>
 
-void saveSystemInfoToFile(const std::string& filename, std::string ComputerName, int osver, std::string WindowsUsername, std::string BuildNum, std::string WinVer,std::string ProcName, std::string TotalMem,std::string gpuName) {
+void saveSystemInfoToFile(const std::string& filename, std::string ComputerName, double osver, std::string WindowsUsername, std::string BuildNum, std::string WinVer,std::string ProcName, std::string TotalMem,std::string gpuName) {
     // Open the output file
     std::ofstream outputFile(filename);
     if (!outputFile) {
